refactor: rework Debug tools
- split debug code to various files (1 per chart, 1 per geom type)
- clarify things belonging to layers or views
- support dynamic creation/hiding of graphs
