+++ index.html
@@ -66,7 +66,8 @@ and open the template in the editor.
                 console.info('Globe initialized');
             });
 
-            new debug.Debug(globeView, viewerDiv);
+            const d = new debug.Debug(globeView, menuGlobe.gui);
+            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.wgs84TileLayer, d);
             window.globeView = globeView;
 </script>
     </body>
