+++ utils/debug/Main.js
@@ -1,2 +1,4 @@
 export { default as Debug } from './Debug';
 export { default as PointCloudDebug } from './PointCloudDebug';
+export { default as createTileDebugUI } from './TileDebug';
+export { default as create3dTilesDebugUI } from './3dTilesDebug';
